Split kickoff event types from calculator
